+++ example/applied_methods_example/artist_methods_example.dart
@@ -22,69 +22,68 @@ void main() async {
 
   final scrobblenaut = Scrobblenaut(lastFM: lastFMAuth);
 
+  final artistInstance = (await scrobblenaut.artist.getInfo(
+    artist: 'RADWIMPS',
+    username: 'nebulino',
+    autoCorrect: true,
+  ));
+
   print('#########################artist.addTags#############################');
 
   // artist.addTags
   print('Result of addTag request: ' +
-      (await scrobblenaut.artist.addTags(artist: 'RADWIMPS', tags: ['anime']))
-          .toString());
+      (await artistInstance.addTags(tags: ['anime'])).toString());
 
   print('#######################artist.getCorrection#########################');
 
   // artist.getCorrection
-  (await scrobblenaut.artist.getCorrection(artist: 'RADWIMPS'))
-      ?.forEach((Artist artist) {
+  (await artistInstance.getCorrection())?.forEach((Artist artist) {
     print('Artist correction: ${artist.name}');
   });
 
   print('#########################artist.getInfo#############################');
 
   // artist.getInfo
-  final artistGetInfo = (await scrobblenaut.artist.getInfo(
-    artist: 'RADWIMPS',
+  final artistGetInfo = (await artistInstance.getInfo(
     username: 'nebulino',
     autoCorrect: true,
   ));
 
-  print(
-      'Artist Info Name: ${artistGetInfo.name} | Artist URL: ${artistGetInfo.url}');
+  print('Artist Info Name: ${artistGetInfo.name} | '
+      'Artist URL: ${artistGetInfo.url}');
 
   print('########################artist.getSimilar###########################');
 
   // artist.getSimilar
-  (await scrobblenaut.artist.getSimilar(artist: 'RADWIMPS'))
-      ?.forEach((Artist artist) {
+  (await artistInstance.getSimilar())?.forEach((Artist artist) {
     print('Similar Artist Name: ${artist.name}');
   });
 
   print('##########################artist.getTags############################');
 
   // artist.getTags
-  (await scrobblenaut.artist.getTags(artist: 'RADWIMPS'))?.forEach((Tag tag) {
+  (await artistInstance.getTags())?.forEach((Tag tag) {
     print('Tag name: ${tag.name} | Tag url: ${tag.url}');
   });
 
   print('#######################artist.getTopAlbums##########################');
 
   // artist.getTopAlbums
-  (await scrobblenaut.artist.getTopAlbums(artist: 'RADWIMPS'))
-      ?.forEach((Album album) {
+  (await artistInstance.getTopAlbums())?.forEach((Album album) {
     print('Album Name: ${album.name} | Album URL: ${album.url}');
   });
 
   print('########################artist.getTopTags###########################');
 
   // artist.getTopTags
-  (await scrobblenaut.artist.getTopTags(artist: 'RADWIMPS'))
-      ?.forEach((Tag tag) {
+  (await artistInstance.getTopTags())?.forEach((Tag tag) {
     print('Tag Name: ${tag.name} | Tag URL: ${tag.url}');
   });
 
   print('####################artist.getTopTracks#############################');
 
   // artist.getTopTracks
-  (await scrobblenaut.artist.getTopTracks(artist: 'RADWIMPS'))
-      ?.forEach((Track track) {
+  (await artistInstance.getTopTracks())?.forEach((Track track) {
     print('Track Name: ${track.name}');
   });
 
@@ -92,15 +91,12 @@ void main() async {
 
   // artist.removeTag
   print('Result of removeTag request: ' +
-      (await scrobblenaut.artist.removeTag(artist: 'RADWIMPS', tag: 'anime'))
-          .toString());
+      (await artistInstance.removeTag(tag: 'anime')).toString());
 
   print('##########################artist.search#############################');
 
   // artist.search
-  (await scrobblenaut.artist.search(artist: 'RADWIMPS'))
-      .artists
-      ?.forEach((Artist artist) {
+  (await artistInstance.search()).artists?.forEach((Artist artist) {
     print('Artist name from search: ${artist.name}');
   });