+++ example/methods_example/scrobble_example.dart
@@ -1,9 +1,9 @@
-import 'package:scrobblenaut/lastfm.dart';
 /**
  * Scrobblenaut - A deadly simple Last.FM API Wrapper for Dart.
  * Copyright (c) 2020 Nebulino
  */
 
+import 'package:scrobblenaut/lastfm.dart';
 import 'package:scrobblenaut/scrobblenaut.dart';
 import 'package:scrobblenaut/scrobblenaut_helpers.dart';