+++ lib/src/extensions/track_extension.dart
@@ -1,0 +1,9 @@
+/**
+ * Scrobblenaut - A deadly simple Last.FM API Wrapper for Dart.
+ * Copyright (c) 2020 Nebulino
+ */
+
+import 'package:scrobblenaut/lastfm.dart';
+
+/// Give the ability of applying [TrackMethods] on [Track].
+extension TrackExtension on Track {}